update
git-svn-id: http://svn.tudos.org/repos/l4linux/trunk@68 98c5fe5d-2a70-4002-a3d4-a65c4b1dd206

+++ include/asm-l4/arch-i386/elf.h
@@ -31,6 +31,6 @@
 #define VDSO_HIGH_BASE (UPAGE_USER_ADDRESS)
 
 #undef VDSO_SYM
-#define VDSO_SYM(x) (x)
+#define VDSO_SYM(x) ((unsigned long)(x))
 
 #endif /* ! __ASM_L4__ARCH_I386__ELF_H__ */
